requirements.txt
django==2.1.7
djangorestframework==3.9.2
gensim==3.7.1
celery==4.2.2
elasticsearch==6.3.1
redis==3.2.1
django-multiselectfield
drf-nested-routers
pandas
scikit-learn
django-rest-auth[with_social]
psutil
matplotlib
termcolor
keras
django-extensions
joblib
